fix minor errors. successfully ran queries and mutations of users, posts, and create user in graphql
